+++ packages/idstool/lib/ids-decomposer.ts
@@ -16,6 +16,11 @@ const encodeMap: Partial<Record<string, string>> = {
     "⿹": "ur",
     "⿺": "ll",
     "⿻": "O",
+    "⿼": "sr",
+    "⿽": "lr",
+    "⿾": "F",
+    "⿿": "R",
+    "㇯": "S",
     "↔": "F",
     "↷": "R",
     "⊖": "S",
@@ -165,25 +170,25 @@ export class IDSDecomposer {
      * Replace all subtractions into IDSs not containing subtractions.
      *
      * examples:
-     * - replace 𠆱=⿰亻⊖斗丶 into 𠆱=⿰亻x and 斗=⿻x丶
-     * - replace 㱐=⊖武㇂ into 武=⿻㱐㇂
+     * - replace 𠆱=⿰亻㇯斗丶 into 𠆱=⿰亻x and 斗=⿻x丶
+     * - replace 㱐=㇯武㇂ into 武=⿻㱐㇂
      *
      * where x stands for a unique token
      */
     private reduceAllSubtractions() {
         function* invert(char: string, subtraction: string[]): Generator<[char: string, tokens: string[]]> {
-            if (subtraction[0] !== "⊖") throw new RangeError("the first token of subtraction is not ⊖")
+            if ((subtraction[0] !== "⊖") && (subtraction[0] !== "㇯")) throw new RangeError("the first token of subtraction is not ⊖")
             if (nodeLength(subtraction, 1) !== 1) throw new Error("unimplemented")
-            // char = ⊖ minuend subtrahend
+            // char = ㇯ minuend subtrahend
             // ≡ minuend = ⿻ char subtrahend
             const minuend = subtraction[1]
             const subtrahend = subtraction.slice(2)
             if (nodeLength(subtrahend, 0) !== subtrahend.length) throw new RangeError("the subtrahend is not complete IDS")
-            if (subtrahend.includes("⊖")) throw new Error("unimplemented")
+            if (subtrahend.includes("⊖") || subtrahend.includes("㇯")) throw new Error("unimplemented")
             yield [minuend, ["⿻", char, ...subtrahend]]
         }
         function* process(char: string, tokens: string[]): Generator<[char: string, tokens: string[]]> {
-            const i = tokens.indexOf("⊖")
+            const i = tokens.findIndex(t => t === "⊖" || t === "㇯")
             if (i < 0) {
                 yield [char, tokens]
                 return
@@ -201,8 +206,8 @@ export class IDSDecomposer {
             [],
             ["UCS", "source", "IDS_tokens"],
             { UCS: string, source: string, IDS_tokens: string }
-        >(`select UCS, source, IDS_tokens from tempids where IDS_tokens glob '*⊖*'`).all()
-        this.db.prepare(`delete from tempids where IDS_tokens glob '*⊖*'`).run()
+        >(`select UCS, source, IDS_tokens from tempids where IDS_tokens glob '*[⊖㇯]*'`).all()
+        this.db.prepare(`delete from tempids where IDS_tokens glob '*[⊖㇯]*'`).run()
         const insert = this.db.prepare<[UCS: string, source: string, IDS_tokens: string], [], {}, void>(`insert into tempids (UCS, source, IDS_tokens) values (?, ?, ?)`)
         for (const { UCS, source, IDS_tokens } of pairs) {
             for (const [char, tokens] of process(UCS, IDS_tokens.split(" "))) {
@@ -280,7 +285,7 @@ export class IDSDecomposer {
             if (tokens.length === 1 && tokens[0] === char) {
                 // atomic component (whose IDS is the character itself)
                 yield tokens
-            } else if (tokens[0] === '⊖') {
+            } else if (tokens[0] === '⊖' || tokens[0] === '㇯') {
                 // treat the char as an atomic component
                 yield [char]
             } else {