+++ packages/idstool/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2023-09-18
+
+### Change
+
+- Support new IDS operators of Unicode 15.1
+
 ## [1.3.0] - 2023-06-29
 
 ### Added